feat. file folder create

+++ core/models/user_repository.go
@@ -6,7 +6,7 @@ type UserRepository struct {
 	gorm.Model
 	Identity           string
 	UserIdentity       string
-	ParentId           int64
+	ParentId           int
 	RepositoryIdentity string
 	Ext                string
 	Name               string
